+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 8.12.8
 ------
 * fix incorrect limits & unit for adaptive gain in trainingChoiceWorld  
+* usability improvements for "Show Training Level" tool
 
 8.12.7
 ------